+++ src/service-worker.js
@@ -1,2 +1,0 @@
-console.log('in was-tutorial service worker');
-importScripts('OneSignalSDKWorker.js');